#393 [refactor] 로깅 관련 패키지 정리
