+++ module-common/src/main/java/com/mile/log/discord/DiscordAppender.java
@@ -6,8 +6,8 @@ import ch.qos.logback.classic.spi.ThrowableProxyUtil;
 import ch.qos.logback.core.UnsynchronizedAppenderBase;
 import com.mile.log.discord.exception.ErrorLogAppenderException;
 import com.mile.log.discord.model.EmbedObject;
-import com.mile.utils.logging.MDCUtil;
-import com.mile.utils.logging.StringUtil;
+import com.mile.log.utils.MDCUtil;
+import com.mile.log.utils.StringUtil;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringEscapeUtils;
