+++ module-common/src/main/java/com/mile/log/discord/DiscordWebHook.java
@@ -9,7 +9,7 @@ import com.mile.log.discord.model.Footer;
 import com.mile.log.discord.model.Image;
 import com.mile.log.discord.model.JsonObject;
 import com.mile.log.discord.model.Thumbnail;
-import com.mile.utils.logging.ApiCallUtil;
+import com.mile.log.utils.ApiCallUtil;
 
 import java.awt.Color;
 import java.io.IOException;
