+++ module-common/src/main/java/com/mile/log/filter/wrapper/CachedBodyRequestWrapper.java
@@ -1,4 +1,4 @@
-package com.mile.filter.wrapper;
+package com.mile.log.filter.wrapper;
 
 import jakarta.servlet.ServletInputStream;
 import jakarta.servlet.http.HttpServletRequest;
